requirements.txt
langchain-community
langchain-openai
python-dotenv
chromadb
pypdf
